@@ -1,4 +1,4 @@
-class CreateFriendlyIdSlugs < ActiveRecord::Migration
+class CreateFriendlyIdSlugs < ActiveRecord::Migration[5.0]
   def change
     create_table :friendly_id_slugs do |t|
       t.string   :slug,           :null => false
