@@ -2,7 +2,9 @@
 #
 # This file contains migration options to ease your Rails 5.0 upgrade.
 #
-# Read the Rails 5.0 release notes for more info on each option.
+# Once upgraded flip defaults one by one to migrate to the new default.
+#
+# Read the Guide for Upgrading Ruby on Rails for more info on each option.
 
 # Enable per-form CSRF tokens. Previous versions had false.
 Rails.application.config.action_controller.per_form_csrf_tokens = true
